HWC: minimize precision loss on DSS scaling
This patch reduces inaccuracy of rounding after sequence of rounding
position and size of element.

(Reworded the explanation)

The rounding of the position has a relation to how width or height is
rounded.

Consider just a line at position 0.5f and width 10.5f. The position of x
is rounded to 1, the width is rounded to 11. The position of the last
pixel in the floating point space is 11f but the integer calculation
would be 1 + 11.

Presumably If this isn't corrected you would see either SGX or GC320
get the calculation right when it was rendering (when the UI is idle)
but see the mismatch when using overlays.

Change-Id: Ica6bdbbd5027a89dd05b3432a35c396df39d5cfc
Signed-off-by: Maksym Kreshchuk <x0191602@ti.com>
Signed-off-by: Tony Lofthouse <a0741364@ti.com>