hwc: Add HDMI mirroring support as per HWC1.1
Since HWC1.1 introduced display specific data structures,
the existing HDMI mirror mode has been reworked to make use of
HWC1.1 data structures. And this is the basis for multi-display
support as well.

This patch also supports HDMI as the default display configuration.

Note#
(1) Since mirroring is the only mode supported in this patch,
HDMI mode is fixed to the mirror region at hotplug or bootup.

Change-Id: I612c5d5f7115558022692e855a568ce6be6f1558
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Conflicts:
	hwc/hwc.c
	hwc/hwc_dev.h