hwc: Remove blitter_composition_t state
Most of the blitter_composition fields are duplicate the fields
in kernel data structure, so by using them we can avoid the
duplication. The num_buffers field represents only part of the
buffers in current composition and is most often used together
with num_buffers from composition_t. Having a single buffer
counter makes more sense and is less prone to errors.

Change-Id: If9fe492c916c77ec6cdae8cc956f7c6ae57c2a8a
Signed-off-by: Mykola Ostrovskyy <mykola@ti.com>