hwc: fix dss pipe assignment logic for single layer usecase
Gallery application creates two layers for full screen video playback.
(1) the background layer with transparent window for the overlaying video
(2) Surfaceview layer for video

But during video playback of video resolution same as the display native
resolution (eg: video of 800x480 on J6 evm), there is only one single
layer from Gallery application during full screen mode.
(1) Surfaceview layer for video

When there is only one layer in the layers stack from Surfaceflinger,
and if that needs either scaling or it is NV-12, we can't assign GFX pipe.
This patch fixed the logic to decide the next available dss pipe
to assign to this single layer.

Change-Id: Id1bee3c4969d6fb56885ee11b11cb9463405611c
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>